generic: Mangle bootloader's kernel arguments
The command-line arguments provided by the boot loader will be
appended to a new device tree property: bootloader-args.
If there is a property "append-rootblock" in DT under /chosen
and a root= option in bootloaders command line it will be parsed
and added to DT bootargs with the form: <append-rootblock>XX.
Only command line ATAG will be processed, the rest of the ATAGs
sent by bootloader will be ignored.
This is usefull in dual boot systems, to get the current root partition
without afecting the rest of the system.

Signed-off-by: Adrian Panella <ianchi74@outlook.com>

This patch has been modified to be mvebu specific. The original patch 
did not pass the bootloader cmdline on if no append-rootblock stanza 
was found, resulting in blank cmdline and failure to boot.

Signed-off-by: Michael Gray <michael.gray@lantisproject.com>